// frontend/src/app/interfaces/tarjeta/DatosTarjeta.ts
export class DatosTarjetaDTO {
  idCard: number;
  vkey: string;
  idPersona: number;
  documento: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  nombres: string;
  celular: string;
  email: string;
  saldo: number;
  picPersonaB64: string;
  picPersona: string;
  fechaRegistro: string;
  fechaVigencia: string;
  fechaVigenciaHasta: string;
  huella: string;
  indice: number;
}